BareConvSwitch
Splits out attention computation from the switch class. This variant
takes in an attention computation (pre-softmax) and applies it against
a list of potential transformations. It is probably a better abstract
representation of how the convolutional switch is supposed to work,
but the old ConvSwitch is being kept around because I'm still working
on some experiments with it.